feat: implement Navbar component and refactor Header usage in layout
